can run deployment end-to-end on new service and existing service, on localhost

+++ app/services/minikube_adapter.rb
@@ -10,14 +10,15 @@ class MinikubeAdapter
     ShellAdapter.exec(cmd)
   end
 
-  def self.configure(config_dir:, environment_slug:, service:)
+  def self.configure(config_dir:, environment_slug:, service:, system_config: {})
     env_vars = ServiceConfigParam.key_value_pairs(
       service.service_config_params
              .where(environment_slug: environment_slug)
              .order(:name)
     )
+
     KubernetesAdapter.set_environment_vars(
-      vars: env_vars,
+      vars: env_vars.merge(system_config),
       service: service,
       config_dir: config_dir,
       environment_slug: environment_slug
@@ -28,12 +29,28 @@ class MinikubeAdapter
   # We want rolling zero-downtime updates, and this is
   # definitely not that
   def self.stop(environment_slug:, service:)
-    KubernetesAdapter.delete_service(
-      name: service.slug,
+    environment = ServiceEnvironment.find(environment_slug)
+    if service_is_running?(service: service, environment_slug: environment_slug)
+      KubernetesAdapter.delete_service(
+        name: service.slug,
+        namespace: environment.namespace,
+        context: environment.kubectl_context
+      )
+    end
+    if deployment_exists?(service: service, environment_slug: environment_slug)
+      delete_deployment(service: service, environment_slug: environment_slug)
+    end
+  end
+
+  def self.delete_deployment(environment_slug:, service:)
+    environment = ServiceEnvironment.find(environment_slug)
+    KubernetesAdapter.delete_deployment(
+      name: KubernetesAdapter.deployment_name(service: service, environment_slug: environment_slug),
       namespace: environment.namespace,
       context: environment.kubectl_context
     )
   end
+
 
   def self.start(environment_slug:, service:, tag:, container_port: 3000, host_port: 8080)
     environment = ServiceEnvironment.find(environment_slug)
@@ -47,25 +64,26 @@ class MinikubeAdapter
       KubernetesAdapter.set_image(
         deployment_name: service.slug,
         container_name: service.slug,
-        image: tag
-      )
-    else
-      KubernetesAdapter.run(
-        tag: tag,
-        name: service.slug,
+        image: tag,
         namespace: environment.namespace,
-        context: environment.kubectl_context,
-        port: container_port
-      )
-
-      KubernetesAdapter.expose_node_port(
-        name: service.slug,
-        namespace: environment.namespace,
-        context: environment.kubectl_context,
-        container_port: container_port,
-        host_port: host_port
+        context: environment.kubectl_context
       )
     end
+    KubernetesAdapter.run(
+      tag: tag,
+      name: service.slug,
+      namespace: environment.namespace,
+      context: environment.kubectl_context,
+      port: container_port
+    )
+
+    KubernetesAdapter.expose_node_port(
+      name: service.slug,
+      namespace: environment.namespace,
+      context: environment.kubectl_context,
+      container_port: container_port,
+      host_port: host_port
+    )
   end
 
   # we don't set up ingress for minikube, we just use node ports
@@ -87,6 +105,16 @@ class MinikubeAdapter
     KubernetesAdapter.exists_in_namespace?(
       name: service.slug,
       type: 'service',
+      namespace: environment.namespace,
+      context: environment.kubectl_context
+    )
+  end
+
+  def self.deployment_exists?(environment_slug:, service:)
+    environment = ServiceEnvironment.find(environment_slug)
+    KubernetesAdapter.exists_in_namespace?(
+      name: service.slug,
+      type: 'deployment',
       namespace: environment.namespace,
       context: environment.kubectl_context
     )
